pathie/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-environ>=0.11.2",
    "django-htmx>=1.26.0",
    "django-leaflet>=0.32.0",
    "django-stubs>=5.2.7",
    "django-tailwind>=4.2.0",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]
